chore: cleaned up Kotlin code; removed unused request parameter
