+++ src/main/app/src/app/formulieren/taken/model/aanvullende-informatie.ts
@@ -183,10 +183,7 @@ export class AanvullendeInformatie extends AbstractTaakFormulier {
       this.zaak.initiatorIdentificatie
     ) {
       this.klantenService
-        .ophalenContactGegevens(
-          this.zaak.initiatorIdentificatieType,
-          this.zaak.initiatorIdentificatie,
-        )
+        .ophalenContactGegevens(this.zaak.initiatorIdentificatie)
         .subscribe((gegevens) => {
           if (gegevens.emailadres) {
             const initiatorToevoegenIcon = new ActionIcon(
