+++ src/main/app/src/app/mail/ontvangstbevestiging/ontvangstbevestiging.component.ts
@@ -116,10 +116,7 @@ export class OntvangstbevestigingComponent implements OnInit {
       this.zaak.initiatorIdentificatie
     ) {
       this.klantenService
-        .ophalenContactGegevens(
-          this.zaak.initiatorIdentificatieType,
-          this.zaak.initiatorIdentificatie,
-        )
+        .ophalenContactGegevens(this.zaak.initiatorIdentificatie)
         .subscribe((gegevens) => {
           if (gegevens.emailadres) {
             const initiatorToevoegenIcon = new ActionIcon(
