+++ src/main/app/src/app/zaken/zaak-afhandelen-dialog/zaak-afhandelen-dialog.component.ts
@@ -91,10 +91,7 @@ export class ZaakAfhandelenDialogComponent implements OnInit, OnDestroy {
       this.data.zaak.initiatorIdentificatie
     ) {
       this.klantenService
-        .ophalenContactGegevens(
-          this.data.zaak.initiatorIdentificatieType,
-          this.data.zaak.initiatorIdentificatie,
-        )
+        .ophalenContactGegevens(this.data.zaak.initiatorIdentificatie)
         .subscribe((gegevens) => {
           if (gegevens.emailadres) {
             this.initiatorEmail = gegevens.emailadres;
