+++ src/main/kotlin/net/atos/zac/app/klant/model/bedrijven/RestListBedrijvenParameters.kt
@@ -4,6 +4,9 @@
  */
 package net.atos.zac.app.klant.model.bedrijven
 
+import net.atos.client.kvk.model.KvkZoekenParameters
+import org.apache.commons.lang3.StringUtils
+
 data class RestListBedrijvenParameters(
     var kvkNummer: String? = null,
     var vestigingsnummer: String? = null,
@@ -13,3 +16,25 @@ data class RestListBedrijvenParameters(
     var huisnummer: Int? = null,
     var type: BedrijfType? = null,
 )
+
+fun RestListBedrijvenParameters.toKvkZoekenParameters(): KvkZoekenParameters {
+    val zoekenParameters = KvkZoekenParameters()
+    if (StringUtils.isNotBlank(this.kvkNummer)) {
+        zoekenParameters.kvkNummer = this.kvkNummer
+    }
+    if (StringUtils.isNotBlank(this.vestigingsnummer)) {
+        zoekenParameters.vestigingsnummer = this.vestigingsnummer
+    }
+    if (StringUtils.isNotBlank(this.rsin)) {
+        zoekenParameters.rsin = this.rsin
+    }
+    if (StringUtils.isNotBlank(this.naam)) {
+        zoekenParameters.naam = this.naam
+    }
+    zoekenParameters.type = this.type?.type
+    if (StringUtils.isNotBlank(this.postcode)) {
+        zoekenParameters.postcode = this.postcode
+    }
+    zoekenParameters.huisnummer = this.huisnummer?.toString()
+    return zoekenParameters
+}
