+++ src/main/kotlin/net/atos/zac/app/klant/model/bedrijven/RestVestigingsprofiel.kt
@@ -4,8 +4,13 @@
  */
 package net.atos.zac.app.klant.model.bedrijven
 
+import net.atos.client.kvk.vestigingsprofiel.model.generated.Vestiging
 import nl.lifely.zac.util.AllOpen
 import nl.lifely.zac.util.NoArgConstructor
+import java.util.Locale
+
+private const val VESTIGINGTYPE_HOOFDVESTIGING = "HOOFDVESTIGING"
+private const val VESTIGINGTYPE_NEVENVESTIGING = "NEVENVESTIGING"
 
 @AllOpen
 @NoArgConstructor
@@ -24,3 +29,39 @@ data class RestVestigingsprofiel(
     var voltijdWerkzamePersonen: Int = 0,
     var website: String? = null
 )
+
+fun Vestiging.toRestVestigingsProfiel() = RestVestigingsprofiel(
+    kvkNummer = this.kvkNummer,
+    vestigingsnummer = this.vestigingsnummer,
+    eersteHandelsnaam = this.eersteHandelsnaam,
+    rsin = this.rsin,
+    totaalWerkzamePersonen = this.totaalWerkzamePersonen,
+    deeltijdWerkzamePersonen = this.deeltijdWerkzamePersonen,
+    voltijdWerkzamePersonen = this.voltijdWerkzamePersonen,
+    commercieleVestiging = this.indCommercieleVestiging?.isIndicatie() ?: false,
+    type = if (this.indHoofdvestiging?.isIndicatie() == true) VESTIGINGTYPE_HOOFDVESTIGING else VESTIGINGTYPE_NEVENVESTIGING,
+    sbiHoofdActiviteit = this.sbiActiviteiten
+        .filter { it.indHoofdactiviteit?.isIndicatie() == true }
+        .map { it.sbiOmschrijving }
+        .firstOrNull(),
+    sbiActiviteiten = this.sbiActiviteiten
+        .filter { it.indHoofdactiviteit?.isIndicatie() == false }
+        .map { it.sbiOmschrijving },
+    adressen = this.adressen
+        .map {
+            RestKlantenAdres(
+                it.type,
+                it.indAfgeschermd?.isIndicatie() ?: false,
+                it.volledigAdres
+            )
+        }
+        .toList(),
+    website = this.websites?.first()
+)
+
+private fun String.isIndicatie(): Boolean =
+    when (this.lowercase(Locale.getDefault())) {
+        "ja" -> true
+        "nee" -> false
+        else -> error("Unexpected value: $this")
+    }
