+++ src/main/kotlin/net/atos/zac/app/klant/model/personen/RestPersoon.kt
@@ -4,10 +4,36 @@
  */
 package net.atos.zac.app.klant.model.personen
 
+import net.atos.client.brp.model.generated.AbstractDatum
+import net.atos.client.brp.model.generated.AbstractVerblijfplaats
+import net.atos.client.brp.model.generated.Adres
+import net.atos.client.brp.model.generated.DatumOnbekend
+import net.atos.client.brp.model.generated.JaarDatum
+import net.atos.client.brp.model.generated.JaarMaandDatum
+import net.atos.client.brp.model.generated.PersonenQueryResponse
+import net.atos.client.brp.model.generated.Persoon
+import net.atos.client.brp.model.generated.PersoonBeperkt
+import net.atos.client.brp.model.generated.RaadpleegMetBurgerservicenummerResponse
+import net.atos.client.brp.model.generated.VerblijfadresBinnenland
+import net.atos.client.brp.model.generated.VerblijfadresBuitenland
+import net.atos.client.brp.model.generated.VerblijfplaatsBuitenland
+import net.atos.client.brp.model.generated.VerblijfplaatsOnbekend
+import net.atos.client.brp.model.generated.VolledigeDatum
+import net.atos.client.brp.model.generated.Waardetabel
+import net.atos.client.brp.model.generated.ZoekMetGeslachtsnaamEnGeboortedatumResponse
+import net.atos.client.brp.model.generated.ZoekMetNaamEnGemeenteVanInschrijvingResponse
+import net.atos.client.brp.model.generated.ZoekMetNummeraanduidingIdentificatieResponse
+import net.atos.client.brp.model.generated.ZoekMetPostcodeEnHuisnummerResponse
+import net.atos.client.brp.model.generated.ZoekMetStraatHuisnummerEnGemeenteVanInschrijvingResponse
 import net.atos.zac.app.klant.model.klant.IdentificatieType
 import net.atos.zac.app.klant.model.klant.RestKlant
+import net.atos.zac.util.StringUtil
+import net.atos.zac.util.StringUtil.ONBEKEND
 import nl.lifely.zac.util.AllOpen
 import nl.lifely.zac.util.NoArgConstructor
+import org.apache.commons.lang3.StringUtils
+import java.util.Locale
+import java.util.Objects
 
 @AllOpen
 @NoArgConstructor
@@ -27,4 +53,101 @@ data class RestPersoon(
     override fun getIdentificatie(): String? {
         return bsn
     }
+}
+
+fun List<Persoon>.toRestPersons(): List<RestPersoon> = this.map { it.toRestPersoon() }
+
+fun List<PersoonBeperkt>.toRestPersonen(): List<RestPersoon> = this.map { it.toRestPerson() }
+
+fun Persoon.toRestPersoon() = RestPersoon(
+    bsn = this.burgerservicenummer,
+    geslacht = this.geslacht?.toDescription(),
+    geboortedatum = this.geboorte?.datum?.toStringRepresentation(),
+    verblijfplaats = this.verblijfplaats?.toStringRepresentation(),
+    naam = this.naam?.volledigeNaam,
+)
+
+fun PersoonBeperkt.toRestPerson() = RestPersoon(
+    bsn = this.burgerservicenummer,
+    geslacht = this.geslacht?.toDescription(),
+    geboortedatum = this.geboorte?.datum?.toStringRepresentation(),
+    naam = this.naam?.volledigeNaam,
+    verblijfplaats = this.adressering?.let {
+        StringUtil.joinNonBlankWith(
+            ", ",
+            it.adresregel1,
+            it.adresregel2,
+            it.adresregel3
+        )
+    }
+)
+
+fun PersonenQueryResponse.toRechtsPersonen(): List<RestPersoon> =
+    when (this) {
+        is RaadpleegMetBurgerservicenummerResponse -> this.personen.toRestPersons()
+        is ZoekMetGeslachtsnaamEnGeboortedatumResponse -> this.personen.toRestPersonen()
+        is ZoekMetNaamEnGemeenteVanInschrijvingResponse -> this.personen.toRestPersonen()
+        is ZoekMetNummeraanduidingIdentificatieResponse -> this.personen.toRestPersonen()
+        is ZoekMetPostcodeEnHuisnummerResponse -> this.personen.toRestPersonen()
+        is ZoekMetStraatHuisnummerEnGemeenteVanInschrijvingResponse -> this.personen.toRestPersonen()
+        else -> emptyList()
+    }
+
+private fun Waardetabel.toDescription(): String =
+    if (StringUtils.isNotBlank(this.omschrijving)) this.omschrijving else this.code
+
+private fun AbstractDatum.toStringRepresentation(): String? =
+    when (this) {
+        is VolledigeDatum -> this.datum.toString()
+        is JaarMaandDatum -> String.format(
+            Locale.getDefault(),
+            "%d2-%d4",
+            this.maand,
+            this.jaar
+        )
+
+        is JaarDatum -> String.format(
+            Locale.getDefault(),
+            "%d4",
+            this.jaar
+        )
+
+        is DatumOnbekend -> ONBEKEND
+        else -> null
+    }
+
+private fun AbstractVerblijfplaats.toStringRepresentation(): String? =
+    when (this) {
+        is Adres -> this.verblijfadres?.toStringRepresentation()
+
+        is VerblijfplaatsBuitenland -> this.verblijfadres?.toStringRepresentation()
+
+        is VerblijfplaatsOnbekend -> ONBEKEND
+        else -> null
+    }
+
+private fun VerblijfadresBinnenland.toStringRepresentation(): String {
+    val adres = StringUtils.replace(
+        StringUtil.joinNonBlankWith(
+            StringUtil.NON_BREAKING_SPACE,
+            this.officieleStraatnaam,
+            Objects.toString(this.huisnummer, null),
+            this.huisnummertoevoeging,
+            this.huisletter
+        ),
+        StringUtils.SPACE,
+        StringUtil.NON_BREAKING_SPACE
+    )
+    val postcode = StringUtils.replace(this.postcode, StringUtils.SPACE, StringUtil.NON_BREAKING_SPACE)
+    val woonplaats = StringUtils.replace(this.woonplaats, StringUtils.SPACE, StringUtil.NON_BREAKING_SPACE)
+    return StringUtil.joinNonBlankWith(", ", adres, postcode, woonplaats)
+}
+
+private fun VerblijfadresBuitenland.toStringRepresentation(): String {
+    return StringUtil.joinNonBlankWith(
+        ", ",
+        this.regel1,
+        this.regel2,
+        this.regel3
+    )
 }
